Add CardDeckController
